+++ pkg/deployment/deployment_pod_sync_test.go
@@ -34,6 +34,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/util/constants"
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/resources"
 )
 
 func TestEnsurePod_Sync_Error(t *testing.T) {
@@ -1304,7 +1305,7 @@ func TestEnsurePod_Sync_Worker(t *testing.T) {
 							},
 							ImagePullPolicy: core.PullIfNotPresent,
 							Lifecycle:       createTestLifecycle(api.ServerGroupSyncMasters),
-							Resources:       k8sutil.ExtractPodAcceptedResourceRequirement(resourcesUnfiltered),
+							Resources:       resources.ExtractPodAcceptedResourceRequirement(resourcesUnfiltered),
 							SecurityContext: securityContext.NewSecurityContext(),
 							VolumeMounts: []core.VolumeMount{
 								k8sutil.LifecycleVolumeMount(),
@@ -1399,7 +1400,7 @@ func TestEnsurePod_Sync_Worker(t *testing.T) {
 							},
 							ImagePullPolicy: core.PullIfNotPresent,
 							Lifecycle:       createTestLifecycle(api.ServerGroupSyncMasters),
-							Resources:       k8sutil.ExtractPodAcceptedResourceRequirement(resourcesUnfiltered),
+							Resources:       resources.ExtractPodAcceptedResourceRequirement(resourcesUnfiltered),
 							SecurityContext: securityContext.NewSecurityContext(),
 							VolumeMounts: []core.VolumeMount{
 								k8sutil.LifecycleVolumeMount(),